Update DumpItGameHelper.js
Fixed a "zero-day" issue where the message was misleading. The intent was to tell the user that the ceremony will be held AFTER the end of the month (which I forgot when I wrote the return for this method). A flag is set in the database which is checked and then reset once the ceremony is over.